pin task
Created an action called pinTask in the index controller and passed it down to task-list component which in turn passes it to its single-task child component.
On click of the pin button, the single-task component calls the action that was passed down to it by its parent task-list component, providing it with the task. The task-list component in turns calls the
pinTask in the index controller which updates its pinnedTask value