Add pin task functionality
I added an action in the index controller that handles the setting of
a pinned task. This action handles logic for pinning, unpinning and
replacing a pinned task. This action is passed down to the single task
component through the task list component and is triggered when
the pin/unpin button is clicked.

I also added an attribute to the task model and I use this to display
or hide the task in the task list. There's probably a better way to do
this though as I don't think this attribute belongs to the model.