Add ToggleComplete component
I created a new component that handles the toggling of task completion.
I then refactored the button for toggling "Done/Undo" from the single task
component to the new component. I created an action in the single task
component to handle the toggling of the 'isComplete' property on tasks.
This action is what is passed down to the ToggleComplete component and
is triggered  by a click event on the "Done/Undo" button.