Create an action that gets a task from the single task component and then creates a pinned task object.
Thought Process:
As per the instructions, since the action needs to send tasks to the index controller. I decided to define the action  here so that I can pass it down to the single task component. That way it will be able to send the task that is pinned to the index controller.