Changes made: Started with passing down the pinTask action that I created in the index.js controller.
I passed it as an argument called pinnedTask to the TaskList component. This way I can be able to pass it to the SingleTask component via the TaskList component.

The next thing I did was to create an onclick event listener for unpin button.
This event listner has an action that mutates the pinnedTask object back to null when the user unpins the task